+++ crates/prover/src/core/backend/gpu/mod.rs
@@ -8,10 +8,11 @@ pub mod m31;
 mod quotients;
 // pub mod packedm31;
 
+use std::ffi::c_void;
 use std::fmt::Debug;
 use std::sync::Arc;
 
-use cudarc::driver::{CudaDevice, ValidAsZeroBits};
+use cudarc::driver::{CudaDevice, DeviceRepr, ValidAsZeroBits};
 // use error::Error;
 use once_cell::sync::Lazy;
 
@@ -39,6 +40,24 @@ impl Load for Device {
         column::load_batch_inverse_ptx(&self);
         circle::load_circle(&self);
         self
+    }
+}
+
+unsafe impl DeviceRepr for M31 {
+    fn as_kernel_param(&self) -> *mut c_void {
+        self as *const Self as *mut c_void
+    }
+}
+
+unsafe impl DeviceRepr for QM31 {
+    fn as_kernel_param(&self) -> *mut c_void {
+        self as *const Self as *mut c_void
+    }
+}
+
+unsafe impl DeviceRepr for &mut QM31 {
+    fn as_kernel_param(&self) -> *mut std::ffi::c_void {
+        self as *const Self as *mut _
     }
 }
